auth

+++ backend/src/controllers/notes.ts
@@ -2,12 +2,18 @@ import { RequestHandler, text } from "express";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
 import NoteModel from '../models/node'
+import { assertIsDefined } from "../util/assertIsDefined";
 
 
 // Get all notes
 export const getNotes: RequestHandler = async (req,res,next) =>{
+
+    const authenticatedUserId = req.session.userId;
+
     try{
-        const notes = await NoteModel.find().exec();
+        
+        assertIsDefined(authenticatedUserId);
+        const notes = await NoteModel.find({userId:authenticatedUserId}).exec();
         res.status(200).json(notes);
     }catch(error){
         next(error )
@@ -19,9 +25,14 @@ export const getNotes: RequestHandler = async (req,res,next) =>{
 
 export const getNote: RequestHandler = async (req,res,next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
     
 
     try{
+
+        assertIsDefined(authenticatedUserId);
+         
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,'Invalid note id');
         }
@@ -30,6 +41,11 @@ export const getNote: RequestHandler = async (req,res,next) => {
         if(!note){
             throw createHttpError(404,'Note not found');
         }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You can't access this note")
+        }
+
         res.status(200).json(note);
     }catch(error){
         next(error )
@@ -44,13 +60,18 @@ interface CreateNoteBody{
 export const createNotes: RequestHandler<unknown,unknown,CreateNoteBody,unknown> = async (req,res,next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try{
+
+        assertIsDefined(authenticatedUserId);
+
         if(!title){
             throw createHttpError(400, 'Note must have a title')
         }
         
         const newNote = await NoteModel.create({
+                userId:authenticatedUserId,
                 title:title,
                 text:text,
         });
@@ -76,8 +97,13 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
+
     
     try{
+
+        assertIsDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,'Invalid note id');
         }
@@ -90,6 +116,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if(!note){
             throw createHttpError(404,'Note not found');
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You can't access this note")
         }
 
         note.title = newTitle;
@@ -109,8 +139,13 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req,res,next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
 
     try{
+
+        assertIsDefined(authenticatedUserId);
+
 
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,'Invalid note id');
@@ -120,6 +155,10 @@ export const deleteNote: RequestHandler = async (req,res,next) => {
 
         if(!note){
             throw createHttpError(404,'Note not found');
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You can't access this note")
         }
 
         await note.deleteOne();
